+++ Gemfile
@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-gem "activerecord", ">= 3.0"
+gem "activerecord", ">= 3.2"
 
 group :development do
   gem 'guard-rspec', require: false
