--- lib/wildsearcher/version.rb
@@ -1,3 +1,0 @@
-module Wildsearcher
-  VERSION = "1.0.4"
-end
